add enable_flaky_tests and skip some flaky tests in gctuner
