+++ pkg/gctuner/finalizer_test.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build enable_flaky_tests
+
 package gctuner
 
 import (
